+++ modules/apps/headless/headless-foundation/headless-foundation-test/src/testIntegration/java/com/liferay/headless/foundation/resource/v1_0/test/BaseEmailResourceTestCase.java
@@ -298,19 +298,19 @@ public abstract class BaseEmailResourceTestCase {
 			return new ArrayList<>(items);
 		}
 
-		public int getItemsPerPage() {
+		public long getItemsPerPage() {
 			return itemsPerPage;
 		}
 
-		public int getLastPageNumber() {
+		public long getLastPageNumber() {
 			return lastPageNumber;
 		}
 
-		public int getPageNumber() {
+		public long getPageNumber() {
 			return pageNumber;
 		}
 
-		public int getTotalCount() {
+		public long getTotalCount() {
 			return totalCount;
 		}
 
@@ -318,16 +318,16 @@ public abstract class BaseEmailResourceTestCase {
 		protected Collection<T> items;
 
 		@JsonProperty("pageSize")
-		protected int itemsPerPage;
+		protected long itemsPerPage;
 
 		@JsonProperty
-		protected int lastPageNumber;
+		protected long lastPageNumber;
 
 		@JsonProperty("page")
-		protected int pageNumber;
+		protected long pageNumber;
 
 		@JsonProperty
-		protected int totalCount;
+		protected long totalCount;
 
 	}
